app/requirements.txt
pollination-streamlit>=0.5.5
pollination-streamlit-io>=0.71.1
honeybee-core
